backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dateutil==2.8.2
gunicorn
flask_sqlalchemy==3.0.5
psycopg2-binary
dotenv
jwt
cd